Add asciiOnly option for UglifyJS2Filter
I am using the select2 library with the uglifyjs2 filter, which automatically converts utf8 escaped characters into literal utf8 characters. This becomes a problem with ie8 which doesn't support utf8 characters in javascript. The workaround is to add the --ascii-only option to uglifyjs, as referenced here:

https://github.com/argerim/select2-rails/issues/63#issuecomment-27560315

So could we add that as an option to the UglifyJS2Filter?

Thanks,
Richard